Put texture on route (#1186)
