+++ namui/skia/src/native/shader.rs
@@ -12,11 +12,11 @@ impl NativeShader {
         static NATIVE_SHADER_MAP: LruCache<Shader, NativeShader, 64> = LruCache::new();
 
         NATIVE_SHADER_MAP.get_or_create(shader, |shader| match shader {
-            Shader::Image { src } => NativeShader {
+            Shader::Image { src, tile_mode } => NativeShader {
                 skia_shader: src
                     .skia_image
                     .to_shader(
-                        Some((TileMode::Clamp.into(), TileMode::Clamp.into())),
+                        Some((tile_mode.x.into(), tile_mode.y.into())),
                         skia_safe::SamplingOptions::new(
                             FilterMode::Linear.into(),
                             MipmapMode::Linear.into(),
